requirements.txt
pillow~=9.1
numpy~=1.22
matplotlib~=3.5
pandas~=1.4
torch~=1.11
torchvision~=0.12
torchinfo~=1.6
tensorflow~=2.8
tensorflow-metal~=0.4
